Override getAttribute in eloquent model, move relation methods to HasRelations
Tests are passing
